backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
aiosqlite==0.20.0
google-genai==1.7.0
python-dotenv==1.0.1
python-multipart==0.0.20
pydantic==2.10.6
aiofiles==24.1.0
httpx==0.28.1
